Cart return visual indicator and improved sound
When carts are returned, the cart type is queued up so that the bell
sound and a text indicator are displayed at a defined rate (currently,
10 Hz). Adjust the cart count game data structure to better support
this.

+++ cart-return/Assets/Scripts/Behaviors/InitializeGameData.cs
@@ -19,6 +19,6 @@ public class InitializeGameData : MonoBehaviour
         GameData.init(_gameConfig,
                       playerCart,
                       playerCart,
-                      (uint)stackSize);
+                      stackSize);
     }
 }
